might as well use the float versions of sine and cosine

+++ MeshOperations/meshop.cpp
@@ -30,22 +30,22 @@ namespace meshop {
 	// could do these with matrices but for set axes i'm pretty sure this is faster
 	void transform_rotate_x(float* data, int len, float angle) {
 		for (int i = 0; i < len; i += meshop::vertex_size) {
-			data[i + 1] = data[i + 1] * cos(angle) - data[i + 2] * sin(angle);
-			data[i + 2] = data[i + 1] * sin(angle) - data[i + 2] * cos(angle);
+			data[i + 1] = data[i + 1] * cosf(angle) - data[i + 2] * sinf(angle);
+			data[i + 2] = data[i + 1] * sinf(angle) - data[i + 2] * cosf(angle);
 		}
 	}
 
 	void transform_rotate_y(float* data, int len, float angle) {
 		for (int i = 0; i < len; i += meshop::vertex_size) {
-			data[i + 0] = data[i + 2] * sin(angle) - data[i + 0] * cos(angle);
-			data[i + 2] = data[i + 2] * cos(angle) - data[i + 0] * sin(angle);
+			data[i + 0] = data[i + 2] * sinf(angle) - data[i + 0] * cosf(angle);
+			data[i + 2] = data[i + 2] * cosf(angle) - data[i + 0] * sinf(angle);
 		}
 	}
 
 	void transform_rotate_z(float* data, int len, float angle) {
 		for (int i = 0; i < len; i += meshop::vertex_size) {
-			data[i + 0] = data[i + 0] * cos(angle) - data[i + 1] * sin(angle);
-			data[i + 1] = data[i + 0] * sin(angle) - data[i + 1] * cos(angle);
+			data[i + 0] = data[i + 0] * cosf(angle) - data[i + 1] * sinf(angle);
+			data[i + 1] = data[i + 0] * sinf(angle) - data[i + 1] * cosf(angle);
 		}
 	}
